engine: Don't use createGuidFromString on literals
Guid.createGuidFromString(String) and NGuid.createGuidFromString(String)
are utility methods that wrap new Guid(String) and return a default
(null or Guid.Empty) in case the string passed to them is null.

This means that there is no point in using them when working with
hard-coded string literals, and a simple call to new Guid("...") would
suffice.

Change-Id: I79780b701c6d156e0cdef983fb1a29d6c21b7b4e
Signed-off-by: Allon Mureinik <amureini@redhat.com>