webadmin: Added sorting infrastructure to SearchableListModel
Added comparator to SearchableListModel, which if set slightly alters
the behaviour of setItems(). If it isn't set, then the behaviour should
not be altered at all.

Changed some dependent code that assumed that getItems() returned a
List, which was a breach of the SearchableListModel interface anyway
(as items is Iterable).

Change-Id: Idadafcb33979d85b7ca332cad717b7c1ceadec4b
Signed-off-by: Lior Vernia <lvernia@redhat.com>