gluster: Host state change event log msg corrected
Corrected the event log message shown when a host status is updated.
Earlier it used to show the message as "Detected new Host <Host Name>.
Host state was set to Up.". This used to be little confusing in case of
already existing host whose status is moved to Up from Non-Operational /
Non-responsive.

The same message is chnaged to "State was set to Up for <Host Name>.".
With the new changes the event log messages in case of a newly added
host would be in the form
"Host <Host Name> was added by admin@internal." followed by
"State was set to Up for <Host Name>".

and in the case of existing host, the only message would be -
"State was set to Up for <Host Name>".

Change-Id: I1e7be78c44afbec88039d2e92d133858c99ffe1e
Bug-Url: https://bugzilla.redhat.com/973629
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>