engine:Start Rebalance Gluster volume command
Rebalance gluster volume is an asynchronous operation started from the console.
Calling this command starts an async task in gluster. Gluster returns a task-id
that is used to monitor status of the task and update the UI.

RebalanceGlusterVolumeCommand extends from GlusterAsyncCommandBase that handles
return of task information from Gluster and updating corresponding step.

Change-Id: Iaae7cf53bd77e9b717b268bcf0de91ba515af40c
Signed-off-by: Sahina Bose <sabose@redhat.com>