core: expand VDS monitoring with cluster emulation modes
This patch follows the wiki on
http://wiki.ovirt.org/Cluster_emulation_modes

Added the logic to match the host supported emulated machine flags
against the config vals or the cluster property.

Added logic to transit the host to NON-OPERATIONAL when it
doesn't match the cluster/config values.

Added logic to set the cluser emulated machine value from the first host
that matches the config values.

Bug-Url: https://bugzilla.redhat.com/927874
Change-Id: I033560bcee25ca4fac7d96606ae4cfbe7013ee09
Signed-off-by: Roy Golan <rgolan@redhat.com>