webadmin: refactored host/cluster general tab mode checks
With the recent enhancements to form related tabs, an empty
line is shown for the form items which are not added to the formpanel.
For example "libvirt version" form item won't be added to formpanel
if the application is not running in virt supported mode. This leaves
an empty line between the previous and next row's form items.

The code is refactored to make use of the setIsAvailable() method
of FormItem instead leaving out the formitem itself.

Change-Id: I1e57336b819ee12b66d5f22589b4b85470e1be5c
Bug-Url: https://bugzilla.redhat.com/973607
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>