core: update storage connection without a domain
Allow updating a storage connection regardless whether
there are storage domains using it or not. Add relevant unitests.

Till now it was allowed to update a connection only
if there were storage domains using it. As part of the agenda
to allow users to manage the storage connections independently,
removing this restriction is necessary,
since connections can be added and edited before a storage domain
started to use it. This will be possible via REST, while in
webadmin UI storage domain and the connection are still presented
and managed as one entity.

Change-Id: I4b72f48d6ae47e3d6d2a50ce0b51693d22840161
Signed-off-by: Alissa Bonas <abonas@redhat.com>