core: remove read lock of template being exported - cont
This is a complementary patch to http://gerrit.ovirt.org/#/c/13109/
which made it possible to create a VM from template, using template
provisioning of type 'clone', while the template is being exported.

This patch makes it possible to create a VM from template, using
template provisioning of type 'thin', while the template is being
exported as well.

The dontCheckTemplateImages field is removed from
VmManagementParameterBase as it was set by all the flows that add VM
besides invoking AddVmCommand directly. now that AddVmCommand doesn't
need to check the template images as well, it becomes redundant.

Instead of checking the template images, the AddVmCommand takes shared
locks for the template and its disks, in a similar way to what was done
in http://gerrit.ovirt.org/#/c/13109/ for AddVmFromTemplateCommand.

Change-Id: I2580bc130e3b9565c8b9f515984c0180a633e3c3
Bug-Url: https://bugzilla.redhat.com/920150
Signed-off-by: Arik Hadas <ahadas@redhat.com>