core: avoid having LOCKED domain on a failed reconstruct.
When the reconstruct command is being executed during its CDA phase
the pool's domains statuses are loaded and are checked for being different
then LOCKED.
getStorageDomain() is being called prior to the CDA check from
getJobMessageProperties(), which means that possibly the loaded domain
would be in status LOCKED.

In case of using compensation through reconstruct, the domain status
might be reverted to LOCKED instead of to it's actual status.

This patch only verifies that the loaded domain within
getStorageDomain() isn't LOCKED, a further behaviour change might be
needed generally.

Change-Id: Icf13f36f30d4cf48e8517444a33bab6359416d37
Signed-off-by: Liron Aravot <laravot@redhat.com>