engine: Fixed validation when moving host to gluster cluster
When moving a host to a gluster cluster, the host can be
added to target cluster only if target cluster is empty
or if there's another up server in target cluster
to perform the gluster command to add to gluster cluster.

The current check was causing NPE while performing
the AddGlusterServer command when there's no up server.
Fixing this with this patch.

Change-Id: Id2f2fbbc5da529932da967fc67ea8b7e34459fbd
Bug-Url: https://bugzilla.redhat.com/976896
Signed-off-by: Sahina Bose <sabose@redhat.com>