core: check no running VMs use storage conn
Check during update connection that if the updated
connection is for block storage, if it's used by lun disks,
that the plugged VMs are down.
Also, failing connect to storage or getting domain stats
is no longer failing the entire command, so updating the
storage connection will not depend on them.

Change-Id: I11fe22760952e1319654647eed661a7814675f64
Signed-off-by: Alissa Bonas <abonas@redhat.com>