utils: enable UT for EngineLocalConfig
EngineLocalConfig is singleton without initialization method, as such we
cannot mock it nor control who in the chain actually call it.

the solution taken is to allow overriding its instance using pre-defined
configuration, so that the default getInstance() method will return
already constructed instance.

to ease usage, a test watcher class was added.

Change-Id: Ie877a056e11a83badd5675cf24ffa48898b5786d
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>