core: Slight improvements to LexoNumericComparator
Added ability to choose case-sensitiveness of comparison, while the
default behaviour is case-insensitive, and added a couple of
tests. Also ceased the opportunity to change the comparison's return
value to always be -1, 0 or 1 (it used to be possible to return other
numbers corresponding to String differences, which could have been
confusing).

Change-Id: If85552b7d07903321d08d55a7d6675c79692f385
Signed-off-by: Lior Vernia <lvernia@redhat.com>