core: Return source string if blank is passed
Patch e0b50e8f4caef63f42d192457f1b8a8d6390a379 Removed EncryptionUtils
and introduced EngineEncryptionUtils.
It also inadvertently introduced a behaviour change in the
encrypt/decrypt methods - returning "" if the given string was blank
instead of the old behaviour of returning null if the original was null.
Note that this change broke the DAO tests

This patch returns the old behaviour in order to allow the DAO tests to
run.

Change-Id: If2882371e84c92b0e087a5db8658512ae7875e52
Signed-off-by: Allon Mureinik <amureini@redhat.com>