webadmin: Sorting Providers in AsyncDataProvider
Noticed that providers weren't sorted in the System Tree, so
AsyncDataProvider seemed like it's as good a place as any to perform
the sorting, as likely all its users would need the items to be
sorted. This made redundant the sorting elsewhere. Also changed
SystemTreeModel to use the AsyncDataProvider method.

Change-Id: I6925aab1d5231256124f4ef045bd7a20bd9c8e20
Signed-off-by: Lior Vernia <lvernia@redhat.com>