engine: Restrict networks visibility
Networks' visibility for users is restricted when a user has a role
on either a Data-Center or in a Cluster which an action group that
allows to viewing the entity children (in context of DC - to query the
networks within that DC, in context of Cluster - to query for the
networks which are assigned to the cluster).

Note that there is no demarcation for networks related action groups
within the role which allows viewing the children.

Change-Id: I04765c346f12a1e9f4295e04d625a72746e58ca7
Signed-off-by: Moti Asayag <masayag@redhat.com>