build: checkstyle to disallow final members
Since GWT RPC does not handle final members properly, there is a need to
implement a mechanism to enforce that no member in this package will be
made final by mistake (unless it's a static final constant), as was
done, e.g., in patch I53d6e5bbd66c4d67ec37326b545b65f086600169.

This patch offers such a mechanism by providing a new checkstyle check
which is only configured to run on the common and compat modules.

Change-Id: I809e7f670cf86b47dbd7ae1fc6f73eff9837a129
Signed-off-by: Allon Mureinik <amureini@redhat.com>