core: Remove final members from common package
Since GWT RPC does not handle final members properly, there is a need to
implement a mechanism to enforce that no member in this package will be
made final by mistake (unless it's a static final constant), as was
done, e.g., in patch I53d6e5bbd66c4d67ec37326b545b65f086600169.

This patch is a preparation for such a mechanism (which will be
presented in the next patch) which removes all the final keywords from
member variables in the common package.

Change-Id: I2caa2e4e79e6237401982cd832164e49dee2d297
Signed-off-by: Allon Mureinik <amureini@redhat.com>