core: Move getOrderedVmGuidsForRunMultipleActions
Moved getOrderedVmGuidsForRunMultipleActions(List<Guid>) from DbFacade
(which it has nothing to do with) to VmStaticDao, where it makes sense.
The tests for the method were moved from DbFacadeDAOTest to
VmStaticDAOTest.

A neat side bonus of this refactoring is getting all the capabilities of
BaseDbFacadeDAO for free, allowing the method's code to be made much
clearer.

Change-Id: Ide46b9ad358a5de2815b8d41df2f86f950c9127c
Signed-off-by: Allon Mureinik <amureini@redhat.com>