core: rv GetEntityPermissionsi[ForUserAndGroups]
Change get_entity_permissions and
get_entity_permissions_for_user_and_groups to use a return value instead
of an output parameter.

Although the main reasoning behind this is standardisation and code
cleanliness, there is an added performance bonus as this change removes
an explicit conversion of the returned uuid to a varchar.

Change-Id: I26d94fc765c927763af6bb0fa6775b4aaced6a28
Signed-off-by: Allon Mureinik <amureini@redhat.com>