core: Allow custom OpenStack Network plugin type
The plugin type should be specifiable by the user, as it will allow
installation of custom hooks that will support this specific plugin type
(it's up to the admin to make sure he puts in the correct value).

The old plugin types can still be sent from the UI, which would cause
the corresponding hook to be triggered.

Change-Id: Ib12a76a069d1d4b5c6cac642bfe6a3da1338a4c5
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>