core: Rename parameter class connectStorageServer.
ConnectStorageServerVDSCommandParameter is used as a parameter class to
connect and disconnect
storage server.
This is quite confusing especially in the ISCSIStorageHelper which calls
a command reflected by a variable (could be connect or disconnect) but
use a parameter class of connectStorageServer.

To make the code looks more readable, the parameter class name
has been changed to StorageServerConnectionManagementVDSParameters.

Change-Id: I84f280bf9abe97fe5c7a4b7514aecf0d29621c3a
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>