core: Improve performance of GUID constructor
The GUID constructor that takes an array of bytes as parameter creates
an intermediate string representation that is then immediately
discarded. This patch changes the constructor to build the GUID without
the intermediate string, borrowing the algorithm from a private
constructor of the UUID class. In my environment this reduces the time
to run the constructor from approx 1500ns to approx 75ns. It should also
have a positive effect in garbage collection, as it doesn't generate the
intermediate string. The overall effect in performance is probably not
that big, but every bit helps.

Change-Id: Ia76c3d9e7ee7594d6d925551320635a20399f11f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>