core: Introducing batch to async task mgr
This patch is first step towards inserting batch
updates/inserts to the async tasks manager system.

The patch intoduces elimination of usage of INSERT SELECT
and fnsplitter in order to insert associations of entities
and async tasks to the database.
Instead, AsyncTaskUtils is changed to insert/update an AsyncTask
object, and a collection of objects the newly introduced
AsyncTaskEntity class - a class that represents the association
between async tasks and entities.

Change-Id: Ia6d8dc6095430c99fdb6c8cd8289c23e270fff9e
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>