core, restapi: Introducing ImportVmFromConfigurationCommand
ImportVmFromConfigurationCommand is used for adding a vm to the engine
with a configuration specified in an ovf file.
The command adds the vm and related devices as specified in the given
configuration and then attempts to attach to the vm the related disks.

*The user can override the vm parameters set in the ovf by explicitly
including overriding value in the REST call (take a look in the example
below).

Usage example:
POST to:
http://server:port/api/vms/
(include Content-Type:application/xml header)

<vm>
<cluster><name>cluster_name</name></cluster>
<initialization><configuration><type>ovf</type>
<data><![CDATA[
]]>
</data></configuration></initialization>
<name>OVERRIDING_NAME</name>
</vm>

*The OVF configuration should be passed within CDATA block or after
escaping to prevent error in the parsing of the request.

Change-Id: I9390bcc11e5f3cb8c8c89ce791aabe63fe0ca341
Signed-off-by: Liron Aravot <laravot@redhat.com>