packaging: setup: Changed stage for temp pgpass
Currently temp pgpass file is created in misc stage.
This causes problems for upgrade, when db validations
must be run in validation stage.

The new implementation moves creation of the temp pgpass
file to validation stage, keeping its name and allowing
other validation steps to do work when the file is available.

Change-Id: Ib9fc158a98cfdf36a1b812a4022298dc6e011d05
Signed-off-by: Alex Lourie <alourie@redhat.com>