core: Fixing NPE in case of endCommand failure
This is a suggestion on how to fix an NPE in
CreateAllSnapshotFromVmCommand that is caused by network error at
CreateSnapshotCommand (the child command).
The network error causes transaction rollback and removal of the
new snapshot entity (and making the old one non active).
I had to create this patch in order to verify previous patches in
this patchset.

Bug-Url: https://bugzilla.redhat.com/982003
Change-Id: I78a617f43bd297db13b7dc9ad6649b67c773c144
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>