core: change RemoveVmFromImportExportParamenters to get VM ID
RemoveVmFromImportExportCommand only needs the ID of the VM to work
with, yet RemoveVmFromImportExportParamenters received VM object. at the
REST API and at the UI we were creating VM object just as a wrapper of
the VM ID to pass to the parameters class. This patch changes the
parameter class to receive ID of VM instead of VM object.

This patch also contains minor refactoring in
RemoveVmFromImportExportParamenters class.

Change-Id: I0fff20548248d3c47087d2d004d285c429a1c081
Signed-off-by: Arik Hadas <ahadas@redhat.com>