core: Remove object_column_white_list from fixtures.xml
This fixture is not needed. This is a metadata table and is
not changing within tests.
It should be created by create_schema.sh and be left as is.
A fix was also inserted to the whitelist sql, to change
sshkeyfingerpint to be lowercase so VdsDAOTest will not break.
The breakage of VdsDAOTest exposed that a non privileged user could
get the ssh key fingerprint.

Change-Id: I6f0429c322f923f0b3c725270aa51c121ca27129
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>