core: move VmCommand#removeVmInSpm method
Move VmCommand#removeVmInSpm method to RemoveVmFromImportExportCommand
class as it is the only place where this method is called from.

The code which is related to disk images removal is extracted from the
execute method of RemoveVmFromImportExportCommand to separate method.

In addition, minor refactoring is made in the canDoAction method of
RemoveVmFromImportExportCommand.

Change-Id: Id3945268925fb5f1541fa29c9ef24e97fc243228
Signed-off-by: Arik Hadas <ahadas@redhat.com>