engine: Don't use ValueObjectMap for PM Options
This patch is part of a series designed to remove the ValueObject*
classes.

In this patch, the main usage of ValueObjectMap, the PM options and
secondary options is removed and replaced with a straight forward
HashMap.
In the several places where Map was replaced with HashMap throughout
this patch is was done in order to supply a serializable type that GWT
can handle.

Change-Id: I9e0bf7aedeff0e0b999921298b34494e5e179a1f
Signed-off-by: Allon Mureinik <amureini@redhat.com>