core: add MemoryUtils#createMemoryStateString method
The implementation of "RAM Snapshots" feature introduced new places
where we need to create the string representation of memory state,
besides HibernateVmCommand. This patch reduce the duplication of code
that creates memory state string (the comma-separated string of IDs)
by introducing a new utility method which is responsible for the
String creation, and is now being called from all the relevant places.

In addition, the changeStorageDomainAndPoolInMemoryVolume method in
MemoryUtils class is renamed to changeStorageDomainAndPoolInMemoryState.

Change-Id: Id1fe579cba2e86d95dc4689c38f195db1ea70eab
Signed-off-by: Arik Hadas <ahadas@redhat.com>