engine: StorageConnections ops won't assume VDS/SP
Storage connection operations have nothing to do with the storage pool,
so that parameter was removed.

In addition, the storage connection operations should be able to work
without an active host, and not validate the connection information.
(If a host is passed, connection information is validated).

This patch decouples storage connection commands from storage domain
commands, and stops assuming the presence of a host and/or a storage
pool.

Change-Id: I628bf2a63f64d3acaa3520ad144e1eef744f1204
Signed-off-by: Allon Mureinik <amureini@redhat.com>