core: prevent removing vm from export domain while it is being imported
While VM is being imported from export domain, there was no kind of lock
that prevented other commands from modifing the VM. So while the import
operation was on progress, the user could remove the VM which is
obviously something we need to prevent.

This patch prevents the scenario above from happening by modify the
ImportVmCommand to take shared lock on the VM at the export domain for
the whole execution of the command, and modify the
RemoveVmFromImportExportCommand to take exclusive lock on the VM which
is about to be removed.

Change-Id: I5c34612f754375a9d15816e9972671e1b13023ad
Bug-Url: https://bugzilla.redhat.com/733917
Signed-off-by: Arik Hadas <ahadas@redhat.com>