engine: Create a vnic profile implicitly when adding network
The patch creates a vnic profile for a newly created network
unless the client requested specifically not to.

The purpose of the patch is to allow same flow of: create
network and use the network by a VMs to function the same
without demanding a user to add another step of creating
the profile.

The created profile will carry the same network name and
will not contain any attributes (no port mirroring, no
custom properties).

Change-Id: Ie8adb307aad864089cbccd5e05d8fc591b6cce60
Signed-off-by: Moti Asayag <masayag@redhat.com>