core: fix memory state removal on import vm rollback
In case import vm operation fails, we rollback the changes and in
particular we remove the memory states of the imported vm from the
data domain.

This patch fix a problem that the memory state images were not removed
by modify the condition that determines if the memory state should be
removed in that case.

The memory states removal was also improved:
if there are no memory states in the snapshots of the imported vm,
the creation of MemoryImageRemover will be skipped.

Change-Id: I98b9908c5332ed66494cc56d826f95314f4f0fee
Signed-off-by: Arik Hadas <ahadas@redhat.com>