frontend: if SpiceReleaseCursorKeys is set SPICE is not opened
If the SpiceReleaseCursorKeys changes to for example
"shift+a" than after clicking the console button the
AsyncDataProvider.getComplexValueFromSpiceRedKeysResource
throws an exception.

The problem was that the getComplexValueFromSpiceRedKeysResource
expects to have all parts of the release cursor keys in the
SpiceRedKeys. In case of "shift+a" it finds "shift" but
does not find "a" and fails.

Fixed by first tying to localize the key (look up in SpiceRedKeys)
but if it does not succeed than returning the key itself.

It is safe to do because the result of getComplexValueFromSpiceRedKeysResource
is used only to show up in the title of the SPICE window and the
SpiceRedKeys contains only the localized parts of it.

Change-Id: Iacbaadf7d6d91c04ec4ec79edc73514bad031714
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>