core: Add external identifier to DbUser
We currently store users and groups in the database using the DbUser
entity, and this entity uses the identifier provided by the LDAP server
as its primary key. This assumes that the identifiers provided by LDAP
servers are unique, even those provided by different LDAP servers, which
isn't correct. This change adds a new "ExternalId" attribute to DbUser
intended to store the identifier used by the LDAP server, so in the
future we will be able to use an unique identifier generated by the
engine.

Change-Id: Iaf81ba87a178f3d275ed172e82b0036635f1a28f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>