packaging: setup: ignore sack cache populate errors
for some reason yum when going to cache only mode tries to download
online repositories, and if fails cause the entire transaction to fail.

more over, when going online, it also tries to download online
resources, and if fails cause the entire transaction to fail.

any network issue during this stage will create an undesired failure. so
just ignore these errors and hope for the best, as this sequence is only
used when transaction is rollbacked.

Change-Id: I6f6cba0071da862addcb438d47ac9401f746ee87
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>