core: Add storage validation when plugging an image.
When a plugging a disk to a VM, the engine should validate that the storage
domain that the disk resides on is active.
There is no reason to call VDSM if it can't connect to the storage.

The validation should be done on plug disk, and also on its derived
operation attachDisk.

Change-Id: Ieeb2522bc5aa280b9b98ef728737aeeaa82d1263
Bug-Url: https://bugzilla.redhat.com/969767
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>