core: Sync VM disks with new guid when importing a VM.
Currently ImportVmCommand use three different kinds of containers when
importing a VM: image id, disk id, and the new disk which is about to be
added to the DB.

The use of those containers has dependency with each other,
and should be reflected in only one container so it will remain in sync.

the following fix, put those parameters into one map, and use this map
when copying a new disk into the Data Center.

Change-Id: I5e25c75b95e08bbb5955b4d96a17baff59336e68
Bug-Url: https://bugzilla.redhat.com/970485
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>