packaging: setup: update temp pgpass for non-default username
Current implementation creates temp pgpass file in the beginning
of the setup/upgrade flow. This causes problems when there are
already DB/username present on the system and not empty. The
setup then generates a new username/DB name, but pgpass is left
with original credentials, which causes external tools to fail.

This new implementation updates the temp pgpass file at the correct
stage so it is correct for external tools execution.

Change-Id: I22c2003ff3ccd9fa3a5ea0554518982d6302f5ba
Signed-off-by: Alex Lourie <alourie@redhat.com>