restapi: add storage domain with existing conn id
Add ability to create a new storage domain for file domains with reference
to an existing storage connection that was previously created.
Also, added a backend validation that the given connection is not used by
other storage domains.
Before this patch, the storage connection was created along with
storage domain creation.

See also:
http://www.ovirt.org/Features/Manage_Storage_Connections#Add_new_file_.28nfs.2Fposix.2Flocal.29_storage_domain_with_existing_storage_connection_.28POST.29

Change-Id: Icb0495be711f80d71ad5334080228faee03d03dc
Signed-off-by: Alissa Bonas <abonas@redhat.com>