engine: Validate NfsTimeo and NfsRetrans
Added @Max and @Min validations on nfsTimeo and nfsRetrans in order to
offer a better value validation instead of just failing the command in
VDSM side if a negative value is given.

Note: This patch just improves the error handling, it does not fix the
underlying issue described in the BZ, which should be handled by the
REST-API layer, or possibly by the JAXB implementation.

Related-to: https://bugzilla.redhat.com/962442
Change-Id: I15d35b2547bb36d17e29c90a05854fe6f897d5d2
Signed-off-by: Allon Mureinik <amureini@redhat.com>