core: Fixed lexo-numeric comparator for big numbers
The lexo-numeric comparator was using Integer class which is very
limited to parse numbers, and would cause an exception if the sequence
of numbers if too long.

Change to BigInteger which is unlimited in size, and has the required
functionality. There doesn't seem to be any adverse effect on the
performance of the comparison.

Change-Id: I52ec8c73e85b2505314a4043206d5883e2a43e37
Bug-Url: https://bugzilla.redhat.com/986979
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>