core: add audit log if snapshot is created without memory
It is possible that the user invokes snapshot with memory creation and
the result will be snapshot without memory, if the VM went down after
the command started. So this patch adds an audit log in order to present
to the user a message that in this case the resulted snapshot doesn't
contain memory as requested.

Change-Id: Id66a5c8d394e1b961d7869538603adb7e6fbaeca
Bug-Url: https://bugzilla.redhat.com/985739
Signed-off-by: Arik Hadas <ahadas@redhat.com>