userportal, webadmin: Don't bundle gwt-dev.jar
The GWT RPC mechanism that we currently use needs some classes that are
part of the gwt-dev.jar file since version 2.5.1. This means that we
need to add that gwt-dev.jar file to the WEB-INF/lib directories of the
applications, and that causes class loading problems because gwt-dev.jar
bundles many third party libraries, including its own copy of the Xerces
XML parser. To avoid this issue we are copying into the frontend project
the sources of the classes needed by the GWT RPC mechanism and not
included in gwt-servlet.jar, currently only
com.google.gwt.core.client.GWTBridge. This way we can set the scope of
the gwt-dev dependency to provided so that it won't be included in the
WEB-INF/lib directories.

Change-Id: I0555b185ea8c52469ebe0779f5287b5bd824a569
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>