core: extract memory images removal on remove vm out of transaction
On remove vm command, the data we remove from the DB is being removed
inside a transaction. The memory images removal, which is a vdsm
operation, is also made inside this trasaction and this is wrong -
vdsm operations should be executed out of transactions.

So this patch extract the call to vdsm for removing the memory images
which is taking place as part of remove vm command to be executed out
of transaction.

Change-Id: I908f8e9fe4e276063aa88357c3cbfe9ae56be501
Bug-Url: https://bugzilla.redhat.com/985410
Signed-off-by: Arik Hadas <ahadas@redhat.com>