webadmin,userportal: change suggest box event handling
The problem was that the ListModelTypeAheadListBox could
be closed only by either selecting one of the values or clicking
somewhere outside of it, but not by clicking inside it again
or clicking to the drop down triangle.

Fixed by disabling the auto-hide feature of the suggest box
and handling this events by hand.

The SuggestBoxFocusHandler now contains a field called "enabled"
which enables-disables the handling by this handler. It is set
from FocusHandlerEnablingMouseHandlers. The order of the events is:
onMouseDown
onBlur
onMouseUp
makeing the onBlur bordered between the mouse events. The reason is
that the blur handler is supposed to hide the suggestions only if
the other ways are not used (e.g.: clicking on the triangle or selecting a
suggestion by clicking on some suggestion).

One other fix is that the EnterIgnoringNativePreviewHandler has been
fixed to process the enter event only once.

Change-Id: Ib73fa0160ea1a80246bb8e9cc9a6052fad63f1fe
Bug-Url: https://bugzilla.redhat.com/987845
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>